+++ CHANGELOG.md
@@ -5,16 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## Unreleased
 
-[unreleased]: https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.3...HEAD
+[*see all changes*](https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.4...HEAD)
 
-## [0.0.3] - 2023-08-29
+## [0.0.4] - 2023-09-12
+[0.0.4]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.4
+
+### Added
+
+- **code**: Render included content as code block
+
+### Changed
+
+- `lang` is deprecated and will be removed in favor of `code`
+
+[*see all changes*](https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.3...v0.0.4)
+
+## [0.0.3] - 2023-09-08
 [0.0.3]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.3
 
 ### Fixed
 
 - **caption**: line numbers are no longer off
+
+[*see all changes*](https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.2...v0.0.3)
 
 ## [0.0.2] - 2023-08-29
 [0.0.2]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.2
@@ -29,7 +44,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `includex` errors are raised by default
 - treat as error when no content is found to include
 
+[*see all changes*](https://github.com/jannismain/mkdocs-macros-includex/compare/v0.0.1...v0.0.2)
+
 ## [0.0.1] - 2023-07-21
 [0.0.1]: https://github.com/jannismain/mkdocs-macros-includex/releases/tag/v0.0.1
 
 Initial release
+
+[*see all changes*](https://github.com/jannismain/mkdocs-macros-includex/compare/f0c8a335f...v0.0.1)